fix: Resolve all issues from comprehensive code review
Critical (P0):
- Fix bind group recreation on window resize in renderer
- Implement LazyLock regex caching in evaluator and hlsl modules
- Add expression length limits (100KB) for DoS protection

High Priority (P1):
- Add Copy trait to RenderState for efficiency
- Remove unnecessary clone in engine per-frame loop
- Add SECURITY.md with security policy
- Add CONTRIBUTING.md with development guidelines
- Add rust-toolchain.toml to pin Rust 1.85
- Add dependabot.yml for weekly dependency updates
- Add deny.toml for cargo-deny configuration

Medium/Low (P2/P3):
- Fix 58+ clippy warnings across workspace
- Fix README naming (milk-* → onedrop-*)
- Fix manual_flatten, needless_range_loop patterns
- Fix approx_constant warnings with std::f64::consts
- Add #[allow(dead_code)] for placeholder modules

All tests passing, clippy clean with -D warnings.

+++ onedrop-engine/tests/e2e_test.rs
@@ -5,8 +5,8 @@
 use onedrop_engine::{EngineConfig, MilkEngine};
 use onedrop_parser::parse_preset;
 use std::fs;
-use std::time::Duration;
 use std::path::PathBuf;
+use std::time::Duration;
 
 /// Helper to find preset path (works from workspace root or onedrop-engine dir)
 fn preset_path(name: &str) -> PathBuf {
@@ -14,13 +14,13 @@ fn preset_path(name: &str) -> PathBuf {
         PathBuf::from(format!("test-presets/{}", name)),
         PathBuf::from(format!("../test-presets/{}", name)),
     ];
-    
+
     for path in candidates {
         if path.exists() {
             return path;
         }
     }
-    
+
     panic!("Preset not found: {}", name);
 }
 
@@ -43,72 +43,87 @@ fn test_e2e_preset_loading() {
 
     // Create engine
     let engine = pollster::block_on(MilkEngine::new(config));
-    assert!(engine.is_ok(), "Failed to create engine: {:?}", engine.err());
+    assert!(
+        engine.is_ok(),
+        "Failed to create engine: {:?}",
+        engine.err()
+    );
 }
 
 #[test]
 fn test_e2e_audio_processing() {
     // Test audio sample processing
     let config = EngineConfig::default();
-    
+
     let mut engine = pollster::block_on(MilkEngine::new(config)).unwrap();
-    
+
     // Simulate audio samples (sine wave @ 440 Hz)
     let sample_rate = 44100.0;
     let duration = 1.0 / 60.0; // One frame @ 60 FPS
     let samples_per_frame = (sample_rate * duration) as usize;
-    
+
     let mut audio_samples = Vec::with_capacity(samples_per_frame);
     for i in 0..samples_per_frame {
         let t = i as f32 / sample_rate;
         let sample = (2.0 * std::f32::consts::PI * 440.0 * t).sin();
         audio_samples.push(sample);
     }
-    
+
     // Update engine with audio (also renders)
-    let delta_time = duration as f32;
+    let delta_time = duration;
     let result = engine.update(&audio_samples, delta_time);
-    
+
     // Should succeed
-    assert!(result.is_ok(), "Failed to update engine: {:?}", result.err());
+    assert!(
+        result.is_ok(),
+        "Failed to update engine: {:?}",
+        result.err()
+    );
 }
 
 #[test]
 fn test_e2e_render_frame() {
     // Test rendering a single frame
     let config = EngineConfig::default();
-    
+
     let mut engine = pollster::block_on(MilkEngine::new(config)).unwrap();
-    
+
     // Simulate audio (silence)
     let audio_samples = vec![0.0; 735]; // ~44100 / 60
     let delta_time = 1.0 / 60.0;
-    
+
     // Update (includes render)
     let result = engine.update(&audio_samples, delta_time);
-    assert!(result.is_ok(), "Failed to update/render: {:?}", result.err());
+    assert!(
+        result.is_ok(),
+        "Failed to update/render: {:?}",
+        result.err()
+    );
 }
 
 #[test]
 fn test_e2e_multiple_frames() {
     // Test rendering multiple frames in sequence
     let config = EngineConfig::default();
-    
+
     let mut engine = pollster::block_on(MilkEngine::new(config)).unwrap();
-    
+
     let audio_samples = vec![0.0; 735];
     let delta_time = 1.0 / 60.0;
-    
+
     // Render 60 frames (1 second)
     for frame in 0..60 {
         let update_result = engine.update(&audio_samples, delta_time);
         assert!(update_result.is_ok(), "Frame {} update failed", frame);
     }
-    
+
     // Verify state progressed
     let state = engine.state();
     assert_eq!(state.frame, 60, "Frame counter should be 60");
-    assert!((state.time - 1.0).abs() < 0.01, "Time should be ~1.0 second");
+    assert!(
+        (state.time - 1.0).abs() < 0.01,
+        "Time should be ~1.0 second"
+    );
 }
 
 #[test]
@@ -149,17 +164,17 @@ fn test_e2e_preset_switching() {
 fn test_e2e_beat_detection() {
     // Test beat detection integration
     let config = EngineConfig::default();
-    
+
     let mut engine = pollster::block_on(MilkEngine::new(config)).unwrap();
-    
+
     // Enable beat detection
     engine.enable_beat_detection();
-    
+
     // Simulate bass-heavy audio
     let sample_rate = 44100.0;
     let duration = 1.0 / 60.0;
     let samples_per_frame = (sample_rate * duration) as usize;
-    
+
     let mut audio_samples = Vec::with_capacity(samples_per_frame);
     for i in 0..samples_per_frame {
         let t = i as f32 / sample_rate;
@@ -167,14 +182,14 @@ fn test_e2e_beat_detection() {
         let sample = (2.0 * std::f32::consts::PI * 60.0 * t).sin() * 2.0; // Loud bass
         audio_samples.push(sample);
     }
-    
+
     // Update with bass-heavy audio
-    let delta_time = duration as f32;
+    let delta_time = duration;
     let result = engine.update(&audio_samples, delta_time);
-    
+
     // Should detect beat and potentially change preset
     assert!(result.is_ok());
-    
+
     // Beat detector should be enabled
     assert!(engine.beat_detector().is_enabled());
 }
@@ -183,35 +198,38 @@ fn test_e2e_beat_detection() {
 fn test_e2e_performance_baseline() {
     // Baseline performance test
     use std::time::Instant;
-    
+
     let config = EngineConfig::default();
-    
+
     let mut engine = pollster::block_on(MilkEngine::new(config)).unwrap();
-    
+
     let audio_samples = vec![0.0; 735];
     let delta_time = 1.0 / 60.0;
-    
+
     // Warm up
     for _ in 0..10 {
         engine.update(&audio_samples, delta_time).unwrap();
     }
-    
+
     // Measure 60 frames
     let start = Instant::now();
     for _ in 0..60 {
         engine.update(&audio_samples, delta_time).unwrap();
     }
     let elapsed = start.elapsed();
-    
+
     // Should be able to render 60 frames in less than 1 second
     // (This is a baseline, actual performance depends on GPU)
     println!("Rendered 60 frames in {:?}", elapsed);
     println!("Average frame time: {:?}", elapsed / 60);
-    
+
     // Very lenient check (10 seconds for 60 frames = 6 FPS minimum)
     // Real target is 60 FPS (1 second for 60 frames)
-    assert!(elapsed < Duration::from_secs(10), 
-            "Performance too slow: {:?} for 60 frames", elapsed);
+    assert!(
+        elapsed < Duration::from_secs(10),
+        "Performance too slow: {:?} for 60 frames",
+        elapsed
+    );
 }
 
 #[test]
@@ -223,17 +241,19 @@ fn test_e2e_with_preset() {
 
     // Load a preset
     let preset_file = preset_path("$$$ Royal - Mashup (151).milk");
-    engine.load_preset(&preset_file).expect("Failed to load preset");
-    
+    engine
+        .load_preset(&preset_file)
+        .expect("Failed to load preset");
+
     // Simulate varied audio
     let sample_rate = 44100.0;
     let duration = 1.0 / 60.0;
     let samples_per_frame = (sample_rate * duration) as usize;
-    
+
     // Render 120 frames (2 seconds) with varying audio
     for frame in 0..120 {
         let mut audio_samples = Vec::with_capacity(samples_per_frame);
-        
+
         for i in 0..samples_per_frame {
             let t = (frame * samples_per_frame + i) as f32 / sample_rate;
             // Mix of bass (60 Hz) and treble (4000 Hz)
@@ -241,11 +261,11 @@ fn test_e2e_with_preset() {
             let treb = (2.0 * std::f32::consts::PI * 4000.0 * t).sin() * 0.3;
             audio_samples.push(bass + treb);
         }
-        
+
         let result = engine.update(&audio_samples, duration);
         assert!(result.is_ok(), "Frame {} failed", frame);
     }
-    
+
     // Verify state
     let state = engine.state();
     assert_eq!(state.frame, 120);
